WIP(evo-animation): implement parent traversal backwards through generations
